% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_filetypes_varnames}
\alias{all_filetypes_varnames}
\title{Names of all variables in all filetypes}
\format{This is a data.frame of 1,009 observations over 2 variables


\describe{
  \item{Var}{This is the name of the variable in a SEER-Medicare file}
  \item{which_file}{This is the file where one might find `Var`}
}}
\usage{
all_filetypes_varnames
}
\description{
Names of all variables in all filetypes
}
\keyword{datasets}
