% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{appendix_recodes}
\alias{appendix_recodes}
\title{Appendix for SEER-Medicare 11/2016 Claims Files (May 31, 2017)}
\format{A data.frame of 2,605 observations over 3 variables


\describe{
  \item{Var}{This is the short name representation of a SEER variable}
  \item{Code}{This corresponds to the representation of an observation of `Var` as seen in ASCII FWF data files}
  \item{Meaning}{This is the interpreted/decoded value of `Code`}
}}
\usage{
appendix_recodes
}
\description{
From: https://healthcaredelivery.cancer.gov/seermedicare/medicare/appendix.versionk.pdf
}
\keyword{datasets}
