% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_differential_prioritization.R
\name{calculate_differential_prioritization}
\alias{calculate_differential_prioritization}
\title{Perform a statistical test for differential prioritization}
\usage{
calculate_differential_prioritization(
  augur1,
  augur2,
  permuted1,
  permuted2,
  n_subsamples = 50,
  n_permutations = 1000
)
}
\arguments{
\item{augur1}{Augur results from condition 1, obtained from
\code{\link{calculate_auc}}}

\item{augur2}{Augur results from condition 2, obtained from
\code{\link{calculate_auc}}}

\item{permuted1}{permuted Augur results from condition 1, obtained from
\code{\link{calculate_auc}} with the argument \code{augur_mode = "permute"}}

\item{permuted2}{permuted Augur results from condition 2, obtained from
\code{\link{calculate_auc}} with the argument \code{augur_mode = "permute"}}

\item{n_subsamples}{the number of subsamples to pool when calculating the
mean AUC for each permtation; defaults to 50}

\item{n_permutations}{the total number of mean AUCs to calculate from a
background distribution}
}
\value{
a data frame containing the following columns:
\enumerate{
\item \code{cell_type}: the cell types in the input dataste
\item \code{auc.x}: the AUC in condition 1
\item \code{auc.y}: the AUC in condition 2
\item \code{delta_auc}: the difference in AUCs between conditions
\item \code{b} number of times an equally large difference in AUCs was
observed in the permuted data
\item \code{m}: total number of permutations performed
\item \code{z}: the z score of the observed delta-AUC, relative to the
null distribution
\item \code{pval}: the permutation p-value for the observed delta-AUC
\item \code{padj}: the BH-corrected p-value
}
}
\description{
Execute a permutation test to identify cell types with statistically
significant differences in AUC between two different rounds of cell type
prioritization (for instance, the response to drugs A and B, as compared
to a common untreated control).
}
