% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2010}
\alias{census2010}
\title{Zeroed-out claims files for testing and demonstration}
\format{Each of these files is of the same format as their respectively-named input SAS statements


\describe{
 \item{dme.txt.gz}{- Durable Medical Equipment claims}
 \item{outsaf.txt.gz}{- Outpatient claims}
 \item{CCflag.txt.gz}{- Chronic Conditions Flags}
 \item{nch.txt.gz}{- Carrier Claims}
 \item{pedsf.txt.gz}{- Patient Entitlement and Diagnosis }
 \item{medpar.txt.gz}{- Part A short-stay, long-stay, and skilled nursing facility bills}
 \item{pdesaf.txt.gz}{- Part D prescription coverage claims}
}}
\source{
\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/dme.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/DME.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/outpat.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/OUTPAT.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/CCflag.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/chronic-conditions-flags.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/nch.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/NCH.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/pedsf.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/aboutdata/PEDSF.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/medpar.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/MEDPAR.pdf}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/pdesaf.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/medicare/partd.pdf}

2010 65+ population census data

These data are meant for age-adjustment in the Medicare population. See the 
script 'augur/data-raw/
}
\usage{
census2010
}
\description{
These were produced by replacing every character of sample claims files 
  with the number 0.
}
\keyword{datasets}
