% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_data.R
\name{create_prediction_array}
\alias{create_prediction_array}
\title{Create Prediction Array}
\usage{
create_prediction_array(data, lag = 5)
}
\arguments{
\item{lag}{}
}
\value{

}
\description{
Create Prediction Array
}
