% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_data.R
\name{descale_prediction_data}
\alias{descale_prediction_data}
\title{Denormalize}
\usage{
descale_prediction_data(x, scalers)
}
\arguments{
\item{max_val}{}
}
\value{

}
\description{
Denormalize
}
