% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dissertation_codes}
\alias{dissertation_codes}
\title{Codes used for my dissertation}
\format{This is a data.frame


\describe{
  \item{Treatment.Type}{This is a brief description of the code}
  \item{Code.Classification}{This is a description of which classification system is used, between ICD-9-CM, ICD-O-3, HCPCS, and maybe others.}
  \item{Code}{This is the `Code.Classification`-encoded representation of `Treatment.Type`.}
}}
\usage{
dissertation_codes
}
\description{
Codes used for my dissertation
}
\keyword{datasets}
