% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matching_codes.R
\name{extract_slice_codes}
\alias{extract_slice_codes}
\title{Read Medicare claims files, return rows that have `pttrn` in `codes_column`}
\usage{
extract_slice_codes(directory, pttrn, infile_df, codes_df, codes_column,
  parallel_map = FALSE, corecount = 2)
}
\arguments{
\item{directory}{This is the directory containing those SEER-Medicare 
claims files to be read}

\item{pttrn}{This pattern will be matched to files contained in `directory`,
these files will be read and should have `infile_df` as their
infile_df.}

\item{infile_df}{This is a `data.frame` whose contents include three variables 
`var`, `positions`, and `widths`, each describing the ASCII 
 file to be read.}

\item{codes_df}{This is a `data.frame` containing a variable 'Code' that 
contains the codes to be extracted}

\item{codes_column}{The column of this filetype that contains relevant codes}

\item{parallel_map}{There's an `mclapply` function used to read and
scan claims files, this switch indicates whether to 
use `mclapply`}

\item{corecount}{The number of cores to use for `mclapply`, if it is used.}
}
\value{
A dataframe as specified by `infile_df` and returned by `read_medicare`
}
\description{
This function will accept a directory containing claims files, and a 
  pattern specifying which filetype (as indicated by the first letters
  in the filename, as provided by seer) to read. Note that this 
  pattern can also be used to specify a single or set of files, in 
  addition to any general claims filename prefix.
}
\examples{

# Set up folder filepaths
outpt_data_path <- 
   system.file("extdata", "outpat.txt.gz", package = "augur")

}
\seealso{
extract_all_diagnoses, extract_all_procedures
}
