% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_filename.R
\name{get_filetype}
\alias{get_filetype}
\title{Get the names of files from their filepaths, 
useful because medicare claims are presented 
with one SAS input file for each type of   
claims file.}
\usage{
get_filetype(filepath)
}
\arguments{
\item{filepath}{This is the path to a SAS infile text document.}
}
\description{
Get the names of files from their filepaths, 
useful because medicare claims are presented 
with one SAS input file for each type of   
claims file.
}
