% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_data.R
\name{get_prediction_data}
\alias{get_prediction_data}
\title{Get Prediction Data}
\usage{
get_prediction_data(asset = "BTC-USD", prediction = 5)
}
\arguments{
\item{prediction}{}
}
\value{

}
\description{
Get Prediction Data
}
