% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{grep_raw}
\alias{grep_raw}
\title{Return the matching regex}
\usage{
grep_raw(x, p, v = FALSE)
}
\arguments{
\item{x}{The thing to examine}

\item{p}{The pattern to grep}

\item{v}{Whether to return non-matching characters}
}
\description{
I suspect there's a better way to do this.
}
