% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infiles}
\alias{infiles}
\title{Fixed-Width File (FWF) Column Specifications}
\format{An object of class \code{list} of length 8.}
\usage{
infiles
}
\description{
For each of the claims and PEDSF files, a SAS input statement is provided to describe which character positions reflect which data. The input statements are available online and linked below, and but they're also included in this package. Refer to "augur/data-raw/infiles.R".
}
\keyword{datasets}
