% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{join_claims}
\alias{join_claims}
\title{Merge a PEDSF df with output from `extract_slice`}
\usage{
join_claims(the_cancers, the_matches)
}
\arguments{
\item{the_cancers}{A PEDSF dataframe}

\item{the_matches}{Output from `extract_slice`}
}
\value{
This function returns the dataframe passed to `the_cancers`, 
        but with three extra columns for each of
}
\description{
This function will accept a PEDSF dataframe and an `extract_slice`-made 
dataframe, returning the original PEDSF dataframe with four extra columns, 
each named with a suffix according to the name of the dataframe passed 
to `the_matches`. The first added column has the prefix `counts_` and 
represents the number of times that `patient_id` was seen in `the_matches`; 
the second column is prefixed `days_`, representing the smallest number 
of days from primary cancer  diagnosis (rounded to the first of the month) 
to the claim in `the_matches`; the fourth added column is prefixed `date_`,
and is the `the_matches`  date closest to primary cancer date of diagnosis.
}
\details{
This function was written for the purpose of implementing filter criteria 
on the basis of Medicare claims data, where the required patient population
must have had a claims diagnosis within some number of days of another 
variable (here, date of primary cancer diagnosis).
}
