% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_filename.R
\name{list_txt_files}
\alias{list_txt_files}
\title{Retrieve file list and types from directory containing
claims data}
\usage{
list_txt_files(files_path)
}
\arguments{
\item{files_path}{The directory containing claims files of interest.}
}
\description{
Retrieve file list and types from directory containing
claims data
}
