% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matching_codes.R
\name{merge_scans}
\alias{merge_scans}
\title{Merge data.frames of logical columns from scanning functions}
\usage{
merge_scans(scan_1, scan_2, code_names)
}
\arguments{
\item{scan_1}{The first scan to merge}

\item{scan_2}{The second scan to merge}

\item{code_names}{Codes to pull, these are essential for names}
}
\description{
This function will merge two of the output from `scan_codes`
}
