% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedsf_utils}
\alias{pedsf_utils}
\title{Dataframes to support recoding SEER and Medicare diagnosis and procedure data}
\format{Four data frames of key-value pairs that can be used with 
       `recode_variable` and appropriate claims files to translate 
        codes into human-readable corresponding drug or procedure names.


\describe{ 
 \item{chemo_drugs}{8703 chemotherapy drugs over 9 recode variables}
 \item{chemo_procs}{569 chemo procedures over 8 recode variables}
 \item{radio_procs}{266 radiotherapy procedures over 6 variables}
 \item{radio_diags}{5 diagnoses described by 5 variables}
}}
\source{
\url{https://crn.cancer.gov/resources/ctcodes-drugs.csv}

\url{https://crn.cancer.gov/resources/ctcodes-procedures.csv}

\url{https://crn.cancer.gov/resources/Radiation Therapy Codes_Procedures.csv}

\url{https://crn.cancer.gov/resources/Radiation Therapy Codes_Diagnoses.csv}

\url{https://healthcaredelivery.cancer.gov/seermedicare/program/inputs/pedsf.txt}

\url{https://healthcaredelivery.cancer.gov/seermedicare/aboutdata/pedsf_attachment_a.pdf}
}
\usage{
pedsf_utils
}
\description{
Each dataframe contains the key-value pairs necessary to recode a variable 
 so that it is interpretable, rather than its coded numeric value. These 
 data come from several sources, the URL for each of which is listed here. 
 For more information on each dataframe listed here, see its respective 
 documentation.
}
\seealso{
recode_variable

PEDSF reading and recoding dataframes

This is a list of three dataframes that are useful for PEDSF data reading and recoding.
  The most important of these three dataframes is `pedsf_format`, whose content 
  was pulled from the SEER-Medicare PDF PEDSF documentation. The 
  `pedsf_recodes` dataframe is, of course, useful in conjunction with the 
  `recode_variable` or `recode_pedsf` functions. Finally, the `pedsf_names`
  dataframe is useful for converting between full versus short variable names.
}
\keyword{datasets}
