% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_prioritization.R
\name{plot_differential_prioritization}
\alias{plot_differential_prioritization}
\title{Plot the results of a differential prioritization analysis}
\usage{
plot_differential_prioritization(results, top_n = 0)
}
\arguments{
\item{results}{the output from
\code{\link{calculate_differential_prioritization}}}

\item{top_n}{optionally, the number of top prioritized cell types to label
in the plot}
}
\value{
a \code{ggplot2} object
}
\description{
After performing a statistical test for differential prioritization using
\code{\link{calculate_differential_prioritization}}, plot the results
in a scatterplot, highlighting cell types with significant differences
between conditions.
}
