% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Plot cell type prioritizations as a 'lollipop' plot}
\usage{
plot_lollipop(augur)
}
\arguments{
\item{augur}{a set of Augur results obtained from \code{\link{calculate_auc}}}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the complete ranked list of prioritized cell types as a 'lollipop' plot
(similar to a bar chart, except with each bar replaced by a point and line).
In addition, the exact value of the mean AUC, to three decimal places, is
printed in the plot for each cell type. For an example, see
\strong{Extended Data Fig. 7b} from
\href{https://dx.doi.org/10.1038/s41587-020-0605-1}{doi:10.1038/s41587-020-0605-1}.
}
