% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterplot.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Compare two cell type prioritizations as a scatterplot}
\usage{
plot_scatterplot(augur1, augur2, top_n = 0)
}
\arguments{
\item{augur1}{first set of Augur results, obtained from
\code{\link{calculate_auc}}}

\item{augur2}{second set of Augur results, obtained from
\code{\link{calculate_auc}}}

\item{top_n}{optionally, add labels for the top n cell types whose AUCs
display the greatest change between the first and second Augur results
into the plot}
}
\value{
a \code{ggplot2} object
}
\description{
Compare two sets of cell type prioritization results, calculated for the
same cell types, by comparing them in a scatterplot, with the AUCs
from the first set of Augur results on the x-axis and the second set on the
y-axis. This function can be useful for understanding how a particular
parameter or preprocessing choice influences cell type prioritization.
}
