% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{read_matches}
\alias{read_matches}
\title{Read a set of `extract_all` CSVs, winnow down columns}
\usage{
read_matches(matches_path, ...)
}
\arguments{
\item{matches_path}{The path to a directory containing `extract_` 
function results saved as CSVs}

\item{...}{Arguments passed to `read_csv`}
}
\description{
Read a set of `extract_all` CSVs, winnow down columns
}
