% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_medicare.R
\name{read_medicare}
\alias{read_medicare}
\title{Read a medicare claims file}
\usage{
read_medicare(data_file, format_df, ...)
}
\arguments{
\item{data_file}{The claims file to be read}

\item{format_df}{The `data.frame` representing data file descriptions provided by SEER-Medicare. This must contain three variables `var`, `positions`, and `widths` corresponding to the variable name, position, and width in the ASCII data files.}

\item{...}{Arguments forwarded to `read_fwf` from `readr`}
}
\description{
Read a medicare claims file
}
