% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matching_codes.R
\name{read_scan}
\alias{read_scan}
\title{Keep rows whose code is present in a specified column}
\usage{
read_scan(filepath, infile_df, codes, codes_column)
}
\arguments{
\item{filepath}{A SEER-Medicare claims file path.}

\item{infile_df}{This is the infile_df used to read the `filepath`. It 
must contain three variables `var`, `positions`, and 
`widths`, each describing the ASCII data file to be read.}

\item{codes}{This is an array of codes that will be searched for in 
`codes_column`}

\item{codes_column}{This is the column of the Medicare claims file data
that will be searched for matching codes}
}
\value{
A `tibble` of the format specified by `infile_df`
}
\description{
Keep rows whose code is present in a specified column
}
