% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_data.R
\name{scale_prediction_data}
\alias{scale_prediction_data}
\title{Scale Prediction Data}
\usage{
scale_prediction_data(data, scalers)
}
\arguments{
\item{scalers}{}
}
\value{

}
\description{
Scale Prediction Data
}
