% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_matching_codes.R
\name{scan_codes}
\alias{scan_codes}
\title{Returns TRUE if a code is present in a string vector, as a dataframe for each code}
\usage{
scan_codes(data_codes, ctp)
}
\arguments{
\item{data_codes}{Actual codes, whitespace between each code}

\item{ctp}{Codes against which to check each diagnosis, codes to pull}
}
\value{
data.frame of TRUE/FALSE for each code, where TRUE indicates that
        the code is present in an element of the data.
}
\description{
Data should be provided as it is returned by `read_medicare`. That is, 
it should be a character array whose contents are of the form:
"00000  00000  00000  00000  00000", in fixed-width format.
}
