% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seer_adj_example}
\alias{seer_adj_example}
\title{SEER Direct Age-Adjusted Rates Example Table}
\format{This is a data.frame with 19 observations of 7 variables.



\describe{
  \item{Age}{}
  \item{Count}{}
  \item{Population}{}
  \item{Crude_Rate}{}
  \item{U.S._2000_Standard_Populations}{}
  \item{Age_Distribution_of_Std_Pop}{}
  \item{Component}{}
}}
\usage{
seer_adj_example
}
\description{
From \url{https://seer.cancer.gov/seerstat/tutorials/aarates/step3.html}, 
this table has weights for the US 2000 Standard Population. See the script 
'augur/data-raw/age-adj.R'.
}
\keyword{datasets}
