% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{select_dates_for_join}
\alias{select_dates_for_join}
\title{Function to select claims dataframe variables for use in `join_claims`}
\usage{
select_dates_for_join(claims_df)
}
\arguments{
\item{claims_df}{The claims dataframe (or `extract_slice` output)}
}
\value{
A dataframe consisting of (patient_id, from_dtd, from_dtm, from_dty),  
       or, in the case of the medpar files, (patient_id, adm_m, adm_d, 
       adm_y, dis_m, dis_d, dis_y)
}
\description{
Function to select claims dataframe variables for use in `join_claims`
}
