% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{split_compound_vars}
\alias{split_compound_vars}
\title{Split and translate columns that are provided as a variable number of observations}
\usage{
split_compound_vars(vr, split_pattern, key_df)
}
\arguments{
\item{vr}{A character array of multiple datapoints to split and translate, where the same number of multiple measurements is present in every row}

\item{split_pattern}{This is the character that signifies the beginning or end of a string that will be translated according to `key_df`}

\item{key_df}{A data.frame containing translations between variable codes and values. See the example, `pedsf_utils$pedsf_recodes`}
}
\description{
Split and translate columns that are provided as a variable number of observations
}
\examples{
#gho_key <- dplyr::filter(pedsf_utils$pedsf_recodes, Var == "gho1")
# implement system file retrieval to make pedsf and example split variable
# pedsf_df <- read_medicare

}
\seealso{
read_medicare
}
