% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Superimpose cell type prioritizations onto a dimensionality reduction plot}
\usage{
plot_umap(
  augur,
  sc,
  mode = c("default", "rank"),
  reduction = "umap",
  palette = "cividis",
  top_n = 0,
  cell_type_col = "cell_type"
)
}
\arguments{
\item{augur}{a set of Augur results obtained from \code{\link{calculate_auc}}}

\item{sc}{a Seurat, monocle3, or SingleCellExperiment object containing
dimensionality reduction coordinates}

\item{mode}{whether to plot the raw AUCs or their relative ranks; one of
\code{'default'} (raw AUCs) or \code{'ranks'}}

\item{reduction}{the type of dimensionality reduction to extract from the
\code{sc} object; defaults to 'umap'}

\item{palette}{the color palette used to visualize the AUC for each cell
type. Can be provided either a string specifying a ColorBrewer (see
\code{\link{scale_color_distiller}}) or viridis (see
\code{\link{scale_color_viridis}}) color palette, or a vector of
colors that will be provided to \code{\link{scale_fill_gradientn}}}

\item{top_n}{optionally, the number of top prioritized cell types to label
in the plot}

\item{cell_type_col}{the column of the \code{meta} data frame, or the
metadata container in the \code{Seurat}/\code{monocle} object, that
contains cell type labels for each cell in the gene-by-cell expression
matrix; defaults to \code{"cell_type"}}
}
\value{
a \code{ggplot} object.
}
\description{
Visualize the global landscape of the perturbation response across a
single-cell dataset by superimposing cell type prioritizations (Augur
AUCs, or their relative rank within the dataset) for each cell type onto
a dimensionality reduction
}
\details{
The following color maps can be specified using the \code{palette} argument:
\enumerate{
\item viridis color scales: magma, inferno, plasma, viridis, or cividis
(the default)
\item sequential ColorBrewer scales: BuGn, BuPu, GnBu, Greens, Greys,
Oranges, OrRd, PuBu, PuBuGn, PuRd, Purples, RdPu, Reds, YlGn, YlGnBu,
YlOrBr, or YlOrRd
\item diverging ColorBrewer scales: BrBG, PiYG, PRGn, PuOr, RdBu, RdGy,
RdYlBu, RdYlGn, or Spectral
\item a vector of colors that will be passed into
\code{scale_color_gradientn}
}
}
